Add tentative GSS styles to stretch an element across the whole page
The element gets pushed to the top of the page, which satisfies the
constraints but does not match my expectations. I would like to keep
the element in place.

Note that the styles are not applied when opening the HTML page locally
(URL starting with file://), only from a web server.